Move exception to same scope where it is thrown
