+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/callback/AttachScannedDocumentsValidator.java
@@ -4,7 +4,6 @@ import com.google.common.collect.Sets;
 import org.slf4j.Logger;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.Documents;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.DuplicateDocsException;
 
 import java.util.List;
 import java.util.Map;
