+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/callback/DuplicateDocsException.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd;
+package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.callback;
 
 public class DuplicateDocsException extends RuntimeException {
     public DuplicateDocsException(String message) {
